remove second mutations file
