+++ src/ducks/HistoricalBalance/Address/AddToWatchlist.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import Button from '@santiment-network/ui/Button'
 import Icon from '@santiment-network/ui/Icon'
-import AddToWatchlistDialog from '../../Watchlists/Actions/Add/Add'
-import { updateWatchlistShort } from '../../Watchlists/gql/mutations'
-import { useAddressWatchlists } from '../../Watchlists/gql/lists/hooks'
 import { Infrastructure } from '../../../utils/address'
+import AddToWatchlistDialog from '../../Watchlists/Actions/Add/Add'
+import { useAddressWatchlists } from '../../Watchlists/gql/lists/hooks'
+import { updateWatchlistShort } from '../../Watchlists/gql/list/mutations'
 import styles from './index.module.scss'
 
 const updateWatchlist = ({ id, listItems }) =>
